Created controller deshboard

+++ src/controllers/HouseController.js
@@ -36,11 +36,11 @@ class HouseController {
     const user = await User.findById(user_id);
     const houses = await House.findById(house_id);
 
-    if (String(user._id) !== String(house.user)) {
+    if (String(user._id) !== String(houses.user)) {
       return res.status(401).json({ error: "Não autorizado" });
     }
 
-    await House.updateOne(  
+    await House.updateOne(
       { _id: house_id },
       {
         user: user_id,
@@ -56,14 +56,21 @@ class HouseController {
   }
 
   async destroy(req, res) {
-
     const { house_id } = req.body;
     const { user_id } = req.headers;
 
-    await House.findByIdAndDelete({_id: house_id});
+    const user = await User.findById(user_id);
+    const houses = await House.findById(house_id);
 
-    return res.json({message: 'excluida com sucesso!'})
+    if (String(user._id) !== String(houses.user)) {
 
+      return res.status(401).json({ error: "Não autorizado" });
+    
+    }
+
+    await House.findByIdAndDelete({ _id: house_id });
+
+    return res.json({ message: "excluida com sucesso!" });
   }
 }
